pyproject.toml
[tool.poetry]
name = "minio-fastapi"
version = "0.1.0"
description = "Demo project. Service for storing and retrieving user files via minio api."
authors = ["Pankrashov Dmitry <admin@dev-io.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
minio = "^7.2.5"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = {extras = ["standard"], version = "^0.28.0"}
pyjwt = "^2.8.0"
dateutils = "^0.6.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
